requirements.txt
Flask==3.0.0
cryptography==41.0.7
pycryptodome==3.19.0
Werkzeug==3.0.1